Reduce the none-inlined while implementation to 1 main class and the true/false constructor classes
This does away with the unneeded node structure and uses the DSL to cache the direct call nodes.
