+++ src/trufflesom/interpreter/nodes/specialized/whileloops/WhileTruePrimitiveNode.java
@@ -1,11 +1,13 @@
 package trufflesom.interpreter.nodes.specialized.whileloops;
 
+import com.oracle.truffle.api.dsl.Cached;
 import com.oracle.truffle.api.dsl.GenerateNodeFactory;
 import com.oracle.truffle.api.dsl.Specialization;
-import com.oracle.truffle.api.frame.VirtualFrame;
+import com.oracle.truffle.api.nodes.DirectCallNode;
 
 import bd.primitives.Primitive;
 import trufflesom.vmobjects.SBlock;
+import trufflesom.vmobjects.SInvokable;
 import trufflesom.vmobjects.SObject;
 
 
@@ -18,9 +20,19 @@ public abstract class WhileTruePrimitiveNode extends WhilePrimitiveNode {
     super(true);
   }
 
-  @Specialization
-  protected SObject doWhileConditionally(final VirtualFrame frame,
-      final SBlock loopCondition, final SBlock loopBody) {
-    return (SObject) whileNode.executeEvaluated(frame, loopCondition, loopBody);
+  @Specialization(limit = "INLINE_CACHE_SIZE",
+      guards = {"loopCondition.getMethod() == cachedLoopCondition",
+          "loopBody.getMethod() == cachedLoopBody"})
+  public final SObject doCached(final SBlock loopCondition, final SBlock loopBody,
+      @Cached("loopCondition.getMethod()") final SInvokable cachedLoopCondition,
+      @Cached("loopBody.getMethod()") final SInvokable cachedLoopBody,
+      @Cached("createCallNode(cachedLoopCondition.getCallTarget())") final DirectCallNode conditionNode,
+      @Cached("createCallNode(cachedLoopBody.getCallTarget())") final DirectCallNode bodyNode) {
+    return doWhileCached(loopCondition, loopBody, conditionNode, bodyNode);
+  }
+
+  @Specialization(replaces = "doCached")
+  public final SObject doUncached(final SBlock loopCondition, final SBlock loopBody) {
+    return doWhileUncachedAndUncounted(loopCondition, loopBody);
   }
 }
